bftw: Don't keep DIR*'s around.
DIR*'s were being kept around so dirfd(dir) could be passed to future
openat() calls.  But DIR*'s are big, holding a cache of filenames etc.
read by readdir().

Instead, store the raw fd and dup() it to open a DIR* with fdopendir().
This way we can call dirclose() as soon as possible, while still keeping
an open fd.

Ideally there would be a way to closedir() without invoking close() on
the underlying fd, but this is a good approximation.

Reduces memory footprint by around 64% in a large directory tree.