parse: Ignore -- on the command line.
find uses -- to indicate the end of the flags.  That is,

    $ find -L -- -type f

is allowed, but

    $ find -- -L -type f

results in an error about an unknown predicate `-L`.  This behaviour
doesn't seem particularly useful -- in particular, unlike most other
tools, it doesn't help you if you want to specify a filename beginning
with a -.

So to ensure bfs is compatible with all GNU find command lines, we just
ignore -- whenever it appears.