bftw: Plug a leak if dirqueue_push() fails
If bftw_add() succeeds but dirqueue_push() fails, we need to clean up
the just-added dircache_entry.  Otherwise it will leak, and we'll also
fail the cache->size == 0 assertion.

Fix it by extracting the dircache-related parts of bftw_pop() into a new
helper function bftw_gc(), and call it from bftw_pop() as well as the
bftw_push() failure path.