exec: Make argument size tracking robust to page-granularity accounting
From looking at the Linux exec() implementation, it seems a big part of
the reason we needed extra headroom was that the arguments/environment
are copied page-by-page, so even a small accounting difference could
result in an error of an entire page size.  Grow the headroom to two
entire pages to account for this.