exec: Make ARG_MAX accounting a bit less restrictive with large pages
For Linux-style accounting, we really only need to handle a single page
of wasted space due to rounding.  Subtracting two pages for extra
headroom was reasonable on systems with 4K pages, but overkill on
systems like ppc64le with 64K pages.  Worse yet was the fact that Alpine
Linux only gives us 128K for arguments.

Instead, only subtract a single page, plus the POSIX-recommended 2048
bytes.

Credit to Mike Sullivan for the initial patch and testing on Alpine
ppc64le.

Fixes:
http://build.alpinelinux.org/buildlogs/build-edge-ppc64le/testing/bfs/bfs-1.2.2-r0.log

Co-authored-by: Mike Sullivan <mksully22@gmail.com>