dstring: Initialize freshly-allocated strings
Previously, a string allocated with dstralloc() had length 0 but no
terminating NUL byte there.  This was problematic if such a string was
used without being modified.

In particular, this was reproducible with bfs -ok by not typing any
response to the prompt.  In that case, uninitialized memory was being
tested for a y/n response, with unpredictable results.