Fix -nouser/-nogroup error handling
The proper way to check for nonexistent users/groups is to set errno to
0 before the get{grg,pwu}id() call, and check it afterwards.

On doing this, it becomes obvious that the call can fail if bftw() is
using all the available FDs, so give them some ephemeral FDs.  It would
be ideal to read the user/group table only once, but this fixes the bug
for now.