fsade: Don't treat EINVAL as an error
acl_get_file() will return EINVAL on some platforms (e.g. FreeBSD,
macOS) when the requested ACL type is not supported for that file.

This fixes `bfs -acl` spewing "Invalid argument." for every file on
FreeBSD with ZFS where ACLs are NFSv4 and thus the POSIX.1e ones are not
supported.