opt: Avoid dangling pointers in de_morgan()
If optimize_{and,or}_expr() relocates expr, we need to update the parent
expr or else we might return garbage.  It seems impossible to actually
trigger this bug right now, since the {and,or} optimizations are
symmetric, but it could be hit if the simplifications in de_morgan()
expose more information than was known previously.