bftw: Make iterative deepening actually do depth-first search
bftw_stream() was always pushing to the end of the queue, resulting in
breadth-first behaviour even when BFTW_DFS was set.  This made iterative
deepening a "worst of both worlds" with the same memory use as BFS, but
much slower due to re-traversals.

Fix it by re-using bftw_batch_{start,finish} from bftw_batch().