Don't call stat() just to determine symbolic lengths
The new bftw_cached_stat() helper gets us stat info if we already have
it, but doesn't call stat() if we don't.  In that case we just take a
guess for the initial length to readlinkat().  This lets us avoid stat()
entirely in many cases for -lname and -printf %l.