tests: Drop capabilities when run as root on Linux
bfs's tests rely on file permissions being enforced, which leads them to
work incorrectly when run as root.  This is probably the most common
packaging issue for bfs, most recently seen with Void Linux's update to
bfs 2.2.

Make it easier on packagers by using capsh, if it's available, to drop
the DAC privileges for the tests.

Link: https://github.com/void-linux/void-packages/pull/29437#issuecomment-798670288
Link: https://salsa.debian.org/lamby/pkg-bfs/-/commit/b173efb35da126adb39b0984219d6a2fd9ff428f