tests/gnu/inum_automount: Work around systemd-mount race
This test used systemd-mount to set up an automount on scratch/mnt.
If a previous test mounted something there, systemd will notice that
asynchronously and may not have noticed the unmount by the time that test
runs.  This led to the test randomly being skipped sometimes when
systemd-mount failed.  It could be reproduced with a loop like this:

    $ for _ in {1..10}; ./tests/tests.sh common/L_mount gnu/inum_automount

Work around it by using scratch/automnt instead to avoid conflicts.