bftw: Add a queue of directories to unwrap
For !BFS_USE_UNWRAPDIR, if a file is still pinned in bftw_closedir(), it
has to stay open until its pincount drops to zero.  Since this happens
in bftw_ioq_pop(), we can't immediately call bftw_unwrapdir() as that
adds to the ioq.  Instead, add it to a list that gets drained by the
next bftw_gc().