bftw: Allow forcing bfs_dir allocation from the main thread
When sorting, we can be forced to pop an unopened directory.  If enough
other directories are already open, that can lead to ENOMEM when we try
to open it synchronously.  To avoid this, force allocations from the
main thread to be attempted even if they would go over the limit.

Also, fix the accounting in bftw_allocdir() if allocation fails.