build: Allow flags.mk to override command line variables
Previously, during something like ./configure LDFLAGS=..., any additions
to LDFLAGS from the generated makefiles were ignored.  I had thought
that sub-make invocations would allow those variables to be overridden,
but that is not the behaviour of make.  So instead, set _LDFLAGS etc.
in the generated files so that they don't conflict.