opt: Don't raise RLIMIT_NOFILE if it would prevent using posix_spawn()
If we raise RLIMIT_NOFILE, we have to lower it before calling exec() for
compatibility with select().  If posix_spawn() doesn't support that, we
fall back to fork(), which is quite a bit slower.

Therefore, if we're going to exec() on most files, it's better to keep
RLIMIT_NOFILE the same to avoid the fork() cost, even though it makes
bftw() somewhat slower.