parse: Handle multiple flags at once like -LEXO2
The POSIX Utility Syntax Guidelines specify that flag groups like -HL
should be handled like -H -L.  GNU find doesn't support grouping flags
in this way, but BSD find does.

To avoid conflicts with non-flag primaries, for now we require at least
one flag in a group to be a capital letter.  That is, we support things
like -Lds but not -ds.  We also do not support -fPATH (without a space)
as it would conflict with -follow, -fprint, etc.  It is impossible to be
compatible with both GNU and BSD find here:

    user@gnu$ find -follow
    link
    link/file
    ...
    user@bsd$ find -follow
    find: ollow: No such file or directory

Link: https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html