xspawn: Test the right macro for posix_spawn_..._addfchdir() support
This should have been benign (just a performance regression), but
FreeBSD has a bug that causes the fork()-based fallback to lock up when
ASAN is enabled, due to fork() and dl_iterate_phdr().  This is not a
complete workaround for that issue, since the fork() fallback may be
used for other reasons.

Fixes: 29ddac2 ("config: Check for posix_spawn_file_actions_addfchdir{,_np}()")