sighook: Return instead of re-raising for faults
This makes the death look exactly like it would if we didn't handle the
signal at all.  Coredumps will point at the right instruction, segfaults
will get logged in dmesg again, etc.

Technically POSIX says this is undefined, but if we get a fault, we've
already done something undefined anyway ;)

Link: https://github.com/rust-lang/rust/issues/26458